hidden/sneaking no longer displays names warrior(sin) multi hit skills no longer use extra dura on weapon mpeater now shows up on the sin that's triggered it, rather then on the person that sees it upon saving accounts it now uses a temp file like with guilds (so it only moves the new files to replace the old once it's finished saving) switched flamingsword to ignore agility on target (lets face it: nothing worse then an x second cooldown skill, that can be easely deflected by sins etc) note: will probably remove part of flamingsword dmg when i get back from work later today, it's doing 200% + 110% dmg instead of  just 210%

+++ Server/MirEnvir/Envir.cs
@@ -376,8 +376,14 @@ namespace Server.MirEnvir
 
             try
             {
-                using (FileStream stream = File.Create(AccountPath))
+                using (FileStream stream = File.Create(AccountPath + "n"))
                     SaveAccounts(stream);
+                if (File.Exists(AccountPath))
+                    File.Move(AccountPath, AccountPath + "o");
+                File.Move(AccountPath + "n", AccountPath);
+                if (File.Exists(AccountPath + "o"))
+                File.Delete(AccountPath + "o");
+
             }
             catch (Exception ex)
             {
@@ -437,9 +443,11 @@ namespace Server.MirEnvir
                 string newfilename = fStream.Name;
                 fStream.EndWrite(result);
                 fStream.Dispose();
-                File.Move(oldfilename, oldfilename + 'o');
+                if (File.Exists(oldfilename))
+                    File.Move(oldfilename, oldfilename + "o");
                 File.Move(newfilename, oldfilename);
-                File.Delete(oldfilename + 'o');
+                if (File.Exists(oldfilename + "o"))
+                    File.Delete(oldfilename + "o");
             }
 
         }
@@ -461,7 +469,7 @@ namespace Server.MirEnvir
                 }
 
                 SaveAccounts(mStream);
-                FileStream fStream = new FileStream(AccountPath, FileMode.Create);
+                FileStream fStream = new FileStream(AccountPath + "n", FileMode.Create);
 
                 byte[] data = mStream.ToArray();
                 fStream.BeginWrite(data, 0, data.Length, EndSaveAccounts, fStream);
@@ -474,8 +482,15 @@ namespace Server.MirEnvir
 
             if (fStream != null)
             {
+                string oldfilename = fStream.Name.Substring(0, fStream.Name.Length - 1);
+                string newfilename = fStream.Name;
                 fStream.EndWrite(result);
                 fStream.Dispose();
+                if (File.Exists(oldfilename))
+                    File.Move(oldfilename, oldfilename + "o");
+                File.Move(newfilename, oldfilename);
+                if (File.Exists(oldfilename + "o"))
+                    File.Delete(oldfilename + "o");
             }
 
             Saving = false;
